minor: refactor: update import paths for GitHubContributionCalendar to use the new types directory
