+++ src/app/[username]/SkylinePage.tsx
@@ -8,7 +8,7 @@ import Skyline3d from "../../components/3d/Skyline3D";
 import { useSearchParams, useRouter } from 'next/navigation';
 import { formatDate, getFirstDayOfYearFromLastDay, structureTimelineByWeek } from "../../utils/generateContributionTimeline";
 import { useEffect, useState } from "react";
-import { GitHubContributionCalendar } from 'src/github-types';
+import { GitHubContributionCalendar } from 'src/types/github-types';
 import { Lights } from "src/components/3d/Lights";
 import { OrbitControls, PerspectiveCamera } from "@react-three/drei";
 import Toggle from "src/components/Toggle";
