+++ src/app/[username]/embed/SkylineEmbed.tsx
@@ -7,7 +7,7 @@ import Skyline3d from "../../../components/3d/Skyline3D";
 import { useSearchParams } from 'next/navigation';
 import { getFirstDayOfYearFromLastDay, structureTimelineByWeek } from "../../../utils/generateContributionTimeline";
 import { useEffect, useState } from "react";
-import { GitHubContributionCalendar } from 'src/github-types';
+import { GitHubContributionCalendar } from 'src/types/github-types';
 import { Lights } from "src/components/3d/Lights";
 import { OrbitControls, PerspectiveCamera } from "@react-three/drei";
